+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to the Vault Manager extension will be documented in this file.
+All notable changes to the Vault extension will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2025-10-16
 
 ### Added
-- Initial release of Vault Manager extension
+- Initial release of Vault extension
 - Embedded vault logic directly in realm_backend canister
 - ckBTC balance tracking and management
 - Transaction history sync from ICRC ledger